requirements.txt
streamlit==0.73.1
PyQRCode==1.2.1
roman==3.3
pyfiglet==0.8.post1
upsidedown==0.4
Pillow==8.1.0
